pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfd-operator"
version = "0.1.0"
description = "Neural operator surrogate engineering project for 2D airfoil CFD acceleration."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.12",
  "pandas>=2.1",
  "matplotlib>=3.8",
  "scikit-learn>=1.4",
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "torch>=2.1",
  "fastapi>=0.110",
  "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
testpaths = [
  "tests",
]
